@@ -18,16 +18,16 @@
     "start:server": "node ./app/server/server.js"
   },
   "dependencies": {
-    "@angular/animations": "13.3.11",
+    "@angular/animations": "14.0.4",
     "@angular/cdk": "13.3.9",
-    "@angular/common": "13.3.11",
-    "@angular/compiler": "13.3.11",
-    "@angular/core": "13.3.11",
-    "@angular/forms": "13.3.11",
+    "@angular/common": "14.0.4",
+    "@angular/compiler": "14.0.4",
+    "@angular/core": "14.0.4",
+    "@angular/forms": "14.0.4",
     "@angular/material": "13.3.9",
-    "@angular/platform-browser": "13.3.11",
-    "@angular/platform-browser-dynamic": "13.3.11",
-    "@angular/router": "13.3.11",
+    "@angular/platform-browser": "14.0.4",
+    "@angular/platform-browser-dynamic": "14.0.4",
+    "@angular/router": "14.0.4",
     "@formatjs/intl-relativetimeformat": "11.0.2",
     "@ngrx/effects": "12.4.0",
     "@ngrx/store": "12.4.0",
@@ -51,34 +51,30 @@
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.8",
+    "@angular-devkit/build-angular": "14.0.4",
     "@angular-eslint/builder": "12.7.0",
     "@angular-eslint/eslint-plugin": "12.7.0",
     "@angular-eslint/eslint-plugin-template": "12.7.0",
     "@angular-eslint/schematics": "12.7.0",
     "@angular-eslint/template-parser": "12.7.0",
-    "@angular/cli": "13.3.8",
-    "@angular/compiler-cli": "13.3.11",
-    "@angular/language-service": "13.3.11",
+    "@angular/cli": "14.0.4",
+    "@angular/compiler-cli": "14.0.4",
+    "@angular/language-service": "14.0.4",
     "@babel/core": "7.15.5",
     "@types/node": "16.11.21",
     "@typescript-eslint/eslint-plugin": "4.28.2",
     "@typescript-eslint/parser": "4.28.2",
     "babel-loader": "8.2.2",
-    "eslint": "7.26.0",
-    "eslint-plugin-import": "2.25.3",
-    "eslint-plugin-jsdoc": "37.0.3",
-    "eslint-plugin-prefer-arrow": "1.2.3",
     "husky": "4.2.5",
     "json-loader": "0.5.7",
     "ts-node": "10.2.1",
     "tslint-config-airbnb": "5.11.2",
-    "typescript": "4.6.4",
+    "typescript": "4.7.4",
     "webpack-cli": "4.8.0"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "yarn lint && yarn spellcheck"
+      "pre-commit": ""
     }
   },
   "engines": {
