@@ -3,7 +3,7 @@ import {
   Inject,
   OnInit,
 } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import {
   MAT_DIALOG_DATA,
   MatDialogRef,
@@ -18,10 +18,10 @@ import { LinkAccountsDialogData } from '../../models/link-accounts-dialog-data.i
 })
 export class LinkAccountsDialogComponent implements OnInit {
 
-  public form: FormGroup;
+  public form: UntypedFormGroup;
 
   constructor(
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     @Inject(MatDialogRef) public dialogRef: MatDialogRef<LinkAccountsDialogComponent>,
     @Inject(MAT_DIALOG_DATA) public data: LinkAccountsDialogData,
   ) { }
