@@ -6,8 +6,8 @@ import {
 } from '@angular/core';
 import {
   AbstractControl,
-  FormArray,
-  FormGroup,
+  UntypedFormArray,
+  UntypedFormGroup,
 } from '@angular/forms';
 import { Profile } from '../../../../../api/src/types/profile/profile';
 import { SailorRole } from '../../../../../api/src/types/sail-manifest/sailor-role';
@@ -19,14 +19,14 @@ import { SailorRole } from '../../../../../api/src/types/sail-manifest/sailor-ro
 })
 export class ManifestEditComponent {
 
-  @Input() form: FormGroup;
+  @Input() form: UntypedFormGroup;
   @Input() manifestFormControlName: string;
   @Output() openProfileDialog: EventEmitter<Profile> = new EventEmitter<Profile>();
 
   public SAILOR_ROLE = SailorRole;
 
   public get manifestControls(): AbstractControl[] {
-    return (this.form.get(this.manifestFormControlName) as FormArray).controls;
+    return (this.form.get(this.manifestFormControlName) as UntypedFormArray).controls;
   }
 
   public showProfileDialog(profile: Profile): void {
