@@ -5,8 +5,8 @@ import {
   OnInit,
 } from '@angular/core';
 import {
-  FormBuilder,
-  FormGroup,
+  UntypedFormBuilder,
+  UntypedFormGroup,
   Validators,
 } from '@angular/forms';
 import { Store } from '@ngrx/store';
@@ -21,14 +21,14 @@ import { BasePageComponent } from '../../../base-page/base-page.component';
 })
 export class EmailPasswordComponent extends BasePageComponent implements OnInit {
 
-  public form: FormGroup;
+  public form: UntypedFormGroup;
   public hide = true;
   public resetPasswordLink = resetPassword.toString();
 
   constructor(
     @Inject(HttpClient) private http: HttpClient,
     @Inject(Store) store: Store<any>,
-    @Inject(FormBuilder) private fb: FormBuilder,
+    @Inject(UntypedFormBuilder) private fb: UntypedFormBuilder,
     @Inject(FirebaseService) private firebaseService: FirebaseService,
   ) {
     super(store);
